+++ benchmarks/src/cuda/gpuwattch-ubench/static_power_modeling/INT_FP_TENSOR_STATIC/INT_FP_TENSOR.cu
@@ -55,7 +55,7 @@ void curandErrCheck_(curandStatus_t stat, const char *file, int line) {
 
 
 
-__global__ void power_microbench(float *data1, float *data2, uint32_t *data3, uint32_t *data4, float *res, int div, uint64_t iterations, half *a, half *b, float *c, int M, int N, int K) {
+__global__ void power_microbench(float *data1, float *data2, uint32_t *data3, uint32_t *data4, float *res, int div, unsigned iterations, half *a, half *b, float *c, int M, int N, int K) {
 
   int gid = blockIdx.x*blockDim.x + threadIdx.x;
   register float s1 = data1[gid];
@@ -83,7 +83,7 @@ __global__ void power_microbench(float *data1, float *data2, uint32_t *data3, ui
   
   //ROI
   #pragma unroll 100
-  for (uint64_t j=0 ; j<iterations ; ++j) {
+  for (unsigned j=0 ; j<iterations ; ++j) {
     if((gid%32)<div){
       asm volatile ("{\t\n"
           "add.f32 %0, %1, %0;\n\t"
@@ -129,7 +129,7 @@ void RandomInit_fp(float* data, int n)
 }
 
 int main(int argc, char** argv){
-  uint64_t iterations;
+  unsigned iterations;
   int blocks;
   int div;
   if (argc != 4){
@@ -137,7 +137,7 @@ int main(int argc, char** argv){
     exit(1);
   }
   else {
-    iterations = atoll(argv[1]);
+    iterations = atoi(argv[1]);
     blocks = atoi(argv[2]);
     div = atoi(argv[3]);
   }
